mgen: Move intel specific code under test/intel/ dir
... to add support for other archs.

Note that, mgen selftest does not work on powerpc yet and thus
running 'make' or 'make test' fails to build mgen as of now.
But it won't harm building of 'numatop' binary.
